[CHANGE] put dev tools in separate global console
