+++ .env
@@ -1,0 +1,2 @@
+APP_DEBUG="0"
+APP_ENV="dev"
