+++ .gitignore
@@ -2,6 +2,7 @@
 .idea/
 composer.lock
 bin/*
+!bin/console
 
 ###> symfony/framework-bundle ###
 /.env.local
