+++ src/Command/Template/hooks/canonicalize_filename.sh
