+++ src/Command/Template/hooks/pre-commit
