+++ src/Command/Template/hooks/pre-commit-phpcs
